springBoot eshop -> changed the model to a domain
