+++ src/main/java/by/tms/eshop/domain/Cart.java
@@ -1,24 +1,31 @@
-package by.tms.eshop.model;
+package by.tms.eshop.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import lombok.experimental.SuperBuilder;
 
 @Entity
-@Table(name = "order_Products")
+@Table(name = "carts")
 @NoArgsConstructor
-@SuperBuilder
 @Getter
-public class OrderProduct extends BaseEntity {
+@Setter
+@EqualsAndHashCode(callSuper = true)
+@SuperBuilder
+public class Cart extends BaseEntity {
 
     @ManyToOne
-    @JoinColumn(name = "order_Id")
-    private Order order;
+    @JoinColumn(name = "user_id")
+    private User user;
     @ManyToOne
     @JoinColumn(name = "product_id")
     private Product product;
+    private boolean cart;
+    private boolean favorite;
+    private Integer count;
 }
