+++ src/main/java/by/tms/eshop/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package by.tms.eshop.repository;
 
-import by.tms.eshop.model.Product;
+import by.tms.eshop.domain.Product;
 
 import java.math.BigDecimal;
 import java.util.List;
