+++ src/main/java/by/tms/eshop/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package by.tms.eshop.repository;
 
-import by.tms.eshop.model.User;
+import by.tms.eshop.domain.User;
 
 import java.util.Optional;
